Fix Arrow build (#34)
The original URL is no longer available. With this change, I am adding all mirrors that are available.

I'm also upgrading the version of the dependency.

Note that ExternalProject_Add tries URLs in order. I wasn't able to find any documentation on how to shuffle a list in CMake. It might be good at some point to shuffle the list so the top mirror doesn't end up getting most of the requests.